BrainDisorder/requirements.txt
Flask
tensorflow-gpu
librosa
matplotlib
scikit-image